Effects GUI and improvements to Event instrument assignments (#43)
* Massive refactor to use instrument ID from event and not have an instrument per track

* Removed redundant elements

* Effects are hooked

* format

* Updated egui

* Added effect panel and how instruments are applied

+++ audio_backend/src/id/mod.rs
@@ -1,4 +1,4 @@
-pub type VoiceId = u64;
-pub type SampleId = u64;
-pub type InstrumentId = u64;
-pub type EffectChainId = u64;
+pub type VoiceId = u32;
+pub type SampleId = u32;
+pub type InstrumentId = u32;
+pub type EffectChainId = u32;
